requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic>=2.0
sqlalchemy==2.0.28
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.1.post1
sentence-transformers==2.2.2
faiss-cpu==1.13.2
langchain==0.1.16
langchain-community==0.0.38
google-generativeai>=0.5.0
deep-translator>=1.11.4
langdetect>=1.0.9
streamlit>=1.32.0
tqdm>=4.66.0